cmus: Version bump to 2.5.0 and various fixes
* cmus-2.3.4 needed media-sound/mpeg4ip for decoding mp4 files. That
package has been nuked here and in upstream as well. libmp4v2 is
required instead. Note that it _can_ still be built with
libmp4v2-1.5.0.1, but that lib is too old, and probably will be nuked
soon too. This is why I suggest removing cmus-2.3.4.

* cmus-2.5.0 requires both libmp4v2 and faad2 for decoding mp4 files. I
would just remove the aac option, but some people might find it useful.

* All releases of cmus thusfar require media/ffmpeg[<2.0.0], otherwise
the compilation fails. A fix has already been commited for scm version,
so I added the patch.

* Added some descriptions for the mpc, tremor and wavpack options, since
they were blank. wav and pulse still are, but I think they are more
self-explanatory.

Change-Id: Ibfb2a0577e78359ee1e8d3cfd81e66fdfe2f0c4e