mkvtoolnix: Do parallel build
With a sufficiently new enough rake (which we have, from the ruby dep
used), we can use the "multitask" mode to enable parallel build in
mkvtoolnix.

In multitask mode (-m), rake behaves similarly to 'make' or the rake
fork 'drake' (a copy of which is bundled with mkvtoolnix) and
automatically determines which jobs can run in parallel. It accepts a -j
option to set the jobs, similar to make.

Change-Id: Iec36b1a3c688427dd9ddf2e63d0437d3a89a30a1
Reviewed-on: https://galileo.mailstation.de/gerrit/5399
Reviewed-by: Jenkins <wk@mailstation.de>
Reviewed-by: Niels Ole Salscheider <niels_ole@salscheider-online.de>