easytag: Disable appstream validation to fix the tests
It broke with appstream-glib 0.7.16 and while it would be possible
to fix the appstream file, it is likely to get out of date with
newer versions of appstream-util (similiar to -Werror and newer
compilers) and we don't really do something with appstream metadata
anyway.
Furthermore it avoids dev-libs/appstream-glib as an automagic test
dependency.